fix: checker error

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,8 +10,8 @@
 int get_bit(unsigned long int n, unsigned int index)
 {
 	double rem = n * 0.5;
-	unsigned int num = n * 0.5;
-	unsigned long int i;
+	unsigned long int num = n * 0.5;
+	unsigned int i;
 	char s[64];
 
 		for (i = 0; rem; i++)
